chore: :art: refactor chat component
