+++ src/hooks/chat/useAudioPlayer.ts
@@ -1,0 +1,78 @@
+import { useState, useRef, useEffect } from 'react';
+import { toast } from 'sonner';
+
+export function useAudioPlayer() {
+  const [isAudioPlaying, setIsAudioPlaying] = useState(false);
+  const audioRef = useRef<HTMLAudioElement | null>(null);
+
+  // Initialize audio element
+  useEffect(() => {
+    if (typeof window !== 'undefined') {
+      audioRef.current = new Audio();
+      audioRef.current.onplay = () => setIsAudioPlaying(true);
+      audioRef.current.onended = () => setIsAudioPlaying(false);
+      audioRef.current.onpause = () => setIsAudioPlaying(false);
+
+      return () => {
+        if (audioRef.current) {
+          audioRef.current.pause();
+          audioRef.current.src = '';
+        }
+      };
+    }
+  }, []);
+
+  const base64ToBlob = (base64: string, mimeType: string) => {
+    const byteCharacters = atob(base64);
+    const byteArrays = [];
+
+    for (let i = 0; i < byteCharacters.length; i += 512) {
+      const slice = byteCharacters.slice(i, i + 512);
+      const byteNumbers = new Array(slice.length);
+      for (let j = 0; j < slice.length; j++) {
+        byteNumbers[j] = slice.charCodeAt(j);
+      }
+      const byteArray = new Uint8Array(byteNumbers);
+      byteArrays.push(byteArray);
+    }
+
+    return new Blob(byteArrays, { type: mimeType });
+  };
+
+  const playAudio = (base64Audio: string) => {
+    if (!audioRef.current) return;
+
+    // Stop any currently playing audio
+    stopAudio();
+
+    const audioBlob = base64ToBlob(base64Audio, 'audio/wav');
+    const audioUrl = URL.createObjectURL(audioBlob);
+
+    audioRef.current.src = audioUrl;
+    audioRef.current.play().catch((err) => {
+      console.error('Error playing audio:', err);
+      toast.error('Could not play audio response');
+    });
+  };
+
+  const stopAudio = () => {
+    if (audioRef.current && !audioRef.current.paused) {
+      audioRef.current.pause();
+      audioRef.current.currentTime = 0;
+      setIsAudioPlaying(false);
+    }
+  };
+
+  const handleUserInteraction = () => {
+    if (isAudioPlaying) {
+      stopAudio();
+    }
+  };
+
+  return {
+    isAudioPlaying,
+    playAudio,
+    stopAudio,
+    handleUserInteraction,
+  };
+}
